refactor: move loaders to their own modules
moves the loaders into their own module with each loader type split out
  makes room for the new dev loader
